Reorganize tests, split out WASMs, improve error reporting.
