+++ stellar-contract-env-host/src/test/invocation.rs
@@ -1,0 +1,254 @@
+use super::wasm_examples::ADD_I32;
+use crate::{
+    im_rc::OrdMap,
+    storage::{AccessType, Footprint, Storage},
+    vm::Vm,
+    xdr::{
+        ContractDataEntry, Hash, LedgerEntry, LedgerEntryData, LedgerEntryExt, LedgerKey,
+        LedgerKeyContractData, ScHostObjErrorCode, ScObject, ScStatic, ScVal, ScVec,
+    },
+    CheckedEnv, Host, HostError, Status, Symbol, Tag,
+};
+
+#[test]
+fn invoke_single_contract_function() -> Result<(), HostError> {
+    use stellar_contract_env_common::xdr::ScVmErrorCode;
+
+    let host = Host::default();
+    let id: Hash = [0; 32].into();
+    let vm = Vm::new(&host, id, ADD_I32)?;
+    let a = 4i32;
+    let b = 7i32;
+    let c = 0x7fffffff_i32;
+    let scvec0: ScVec = host.test_scvec::<i32>(&[a, b])?;
+    let res = vm.invoke_function(&host, "add", &scvec0)?;
+    match res {
+        ScVal::I32(v) => assert_eq!(v, a + b),
+        _ => panic!("Wrong result type"),
+    }
+    // overflow
+    let scvec0: ScVec = host.test_scvec::<i32>(&[a, c])?;
+    let res = vm.invoke_function(&host, "add", &scvec0);
+    let code = ScVmErrorCode::TrapUnreachable;
+    assert!(HostError::result_matches_err_status(res, code));
+    Ok(())
+}
+
+#[test]
+fn invoke_cross_contract() -> Result<(), HostError> {
+    let contract_id: Hash = [0; 32].into();
+    let key = ScVal::Static(ScStatic::LedgerKeyContractCodeWasm);
+    let storage_key = LedgerKey::ContractData(LedgerKeyContractData {
+        contract_id: contract_id.clone(),
+        key: key.clone(),
+    });
+    // We unwrap here rather than host.map_err because the host doesn't exist yet.
+    let scob = ScObject::Binary(ADD_I32.try_into().unwrap());
+    let val = ScVal::Object(Some(scob));
+    let le = LedgerEntry {
+        last_modified_ledger_seq: 0,
+        data: LedgerEntryData::ContractData(ContractDataEntry {
+            contract_id,
+            key,
+            val,
+        }),
+        ext: LedgerEntryExt::V0,
+    };
+    let map = OrdMap::unit(storage_key.clone(), Some(le));
+    let mut footprint = Footprint::default();
+    footprint.record_access(&storage_key, AccessType::ReadOnly);
+
+    // initialize storage and host
+    let storage = Storage::with_enforcing_footprint_and_map(footprint, map);
+    let host = Host::with_storage(storage);
+    // create a dummy contract obj as the caller
+    let obj = host.test_bin_obj(&[0; 32])?;
+    // prepare arguments
+    let sym = Symbol::from_str("add");
+    let args = host.test_vec_obj::<i32>(&[1, 2])?;
+
+    let res = host.call(obj.to_object(), sym.into(), args.into())?;
+    assert!(res.is::<i32>());
+    assert!(res.get_tag() == Tag::I32);
+    let i: i32 = res.try_into()?;
+    assert_eq!(i, 3);
+    Ok(())
+}
+
+#[ignore]
+#[test]
+fn invoke_cross_contract_with_err() -> Result<(), HostError> {
+    let contract_id: Hash = [0; 32].into();
+    let key = ScVal::Static(ScStatic::LedgerKeyContractCodeWasm);
+    let storage_key = LedgerKey::ContractData(LedgerKeyContractData {
+        contract_id: contract_id.clone(),
+        key: key.clone(),
+    });
+
+    let code: [u8; 170] = [
+        0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00, 0x01, 0x11, 0x03, 0x60, 0x00, 0x01, 0x7e,
+        0x60, 0x03, 0x7e, 0x7e, 0x7e, 0x01, 0x7e, 0x60, 0x01, 0x7e, 0x01, 0x7e, 0x02, 0x0d, 0x02,
+        0x01, 0x76, 0x01, 0x5f, 0x00, 0x00, 0x01, 0x76, 0x01, 0x38, 0x00, 0x01, 0x03, 0x02, 0x01,
+        0x02, 0x05, 0x03, 0x01, 0x00, 0x10, 0x06, 0x19, 0x03, 0x7f, 0x01, 0x41, 0x80, 0x80, 0xc0,
+        0x00, 0x0b, 0x7f, 0x00, 0x41, 0x80, 0x80, 0xc0, 0x00, 0x0b, 0x7f, 0x00, 0x41, 0x80, 0x80,
+        0xc0, 0x00, 0x0b, 0x07, 0x2f, 0x04, 0x06, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x02, 0x00,
+        0x07, 0x76, 0x65, 0x63, 0x5f, 0x65, 0x72, 0x72, 0x00, 0x02, 0x0a, 0x5f, 0x5f, 0x64, 0x61,
+        0x74, 0x61, 0x5f, 0x65, 0x6e, 0x64, 0x03, 0x01, 0x0b, 0x5f, 0x5f, 0x68, 0x65, 0x61, 0x70,
+        0x5f, 0x62, 0x61, 0x73, 0x65, 0x03, 0x02, 0x0a, 0x29, 0x01, 0x27, 0x01, 0x01, 0x7f, 0x23,
+        0x00, 0x41, 0x10, 0x6b, 0x22, 0x01, 0x24, 0x00, 0x20, 0x01, 0x10, 0x00, 0x42, 0xd1, 0x00,
+        0x20, 0x00, 0x10, 0x01, 0x37, 0x03, 0x00, 0x20, 0x01, 0x29, 0x03, 0x00, 0x20, 0x01, 0x41,
+        0x10, 0x6a, 0x24, 0x00, 0x0b,
+    ];
+    // We unwrap here rather than host.map_err because the host doesn't exist yet.
+    let scob = ScObject::Binary(code.try_into().unwrap());
+    let val = ScVal::Object(Some(scob));
+    let le = LedgerEntry {
+        last_modified_ledger_seq: 0,
+        data: LedgerEntryData::ContractData(ContractDataEntry {
+            contract_id,
+            key,
+            val,
+        }),
+        ext: LedgerEntryExt::V0,
+    };
+    let map = OrdMap::unit(storage_key.clone(), Some(le));
+
+    let mut footprint = Footprint::default();
+    footprint.record_access(&storage_key, AccessType::ReadOnly);
+
+    // initialize storage and host
+    let storage = Storage::with_enforcing_footprint_and_map(footprint, map);
+    let host = Host::with_storage(storage);
+    // create a dummy contract obj as the caller
+    let obj = host.test_bin_obj(&[0; 32])?;
+    // prepare arguments
+    let sym = Symbol::from_str("vec_err");
+    let args = host.test_vec_obj::<i32>(&[1])?;
+
+    // try_call
+    let sv = host.try_call(obj.to_object(), sym.into(), args.clone().into())?;
+    let code = ScHostObjErrorCode::VecIndexOutOfBound;
+    let exp_st: Status = code.into();
+    assert_eq!(sv.get_payload(), exp_st.to_raw().get_payload());
+
+    // call
+    let res = host.call(obj.to_object(), sym.into(), args.into());
+    assert!(HostError::result_matches_err_status(res, code));
+
+    Ok(())
+}
+
+#[ignore]
+#[test]
+fn invoke_cross_contract_lvl2_nested_with_err() -> Result<(), HostError> {
+    // 1st level, the calling contract
+    let id0: Hash = [0; 32].into();
+    let key = ScVal::Static(ScStatic::LedgerKeyContractCodeWasm);
+    let storage_key0 = LedgerKey::ContractData(LedgerKeyContractData {
+        contract_id: id0.clone(),
+        key: key.clone(),
+    });
+    let code0: [u8; 497] = [
+        0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00, 0x01, 0x17, 0x04, 0x60, 0x00, 0x01, 0x7e,
+        0x60, 0x02, 0x7e, 0x7e, 0x01, 0x7e, 0x60, 0x03, 0x7e, 0x7e, 0x7e, 0x01, 0x7e, 0x60, 0x01,
+        0x7e, 0x01, 0x7e, 0x02, 0x1f, 0x05, 0x01, 0x76, 0x01, 0x5f, 0x00, 0x00, 0x01, 0x76, 0x01,
+        0x34, 0x00, 0x01, 0x01, 0x63, 0x01, 0x5f, 0x00, 0x02, 0x01, 0x62, 0x01, 0x33, 0x00, 0x00,
+        0x01, 0x62, 0x01, 0x38, 0x00, 0x01, 0x03, 0x02, 0x01, 0x03, 0x05, 0x03, 0x01, 0x00, 0x11,
+        0x06, 0x19, 0x03, 0x7f, 0x01, 0x41, 0x80, 0x80, 0xc0, 0x00, 0x0b, 0x7f, 0x00, 0x41, 0x87,
+        0x80, 0xc0, 0x00, 0x0b, 0x7f, 0x00, 0x41, 0x90, 0x80, 0xc0, 0x00, 0x0b, 0x07, 0x30, 0x04,
+        0x06, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x02, 0x00, 0x08, 0x64, 0x65, 0x6c, 0x5f, 0x63,
+        0x61, 0x6c, 0x6c, 0x00, 0x05, 0x0a, 0x5f, 0x5f, 0x64, 0x61, 0x74, 0x61, 0x5f, 0x65, 0x6e,
+        0x64, 0x03, 0x01, 0x0b, 0x5f, 0x5f, 0x68, 0x65, 0x61, 0x70, 0x5f, 0x62, 0x61, 0x73, 0x65,
+        0x03, 0x02, 0x0a, 0xc4, 0x02, 0x01, 0xc1, 0x02, 0x02, 0x02, 0x7e, 0x04, 0x7f, 0x23, 0x00,
+        0x41, 0x10, 0x6b, 0x22, 0x06, 0x24, 0x00, 0x41, 0x20, 0x21, 0x03, 0x10, 0x03, 0x21, 0x01,
+        0x03, 0x7e, 0x20, 0x03, 0x04, 0x7e, 0x20, 0x03, 0x41, 0x01, 0x6b, 0x21, 0x03, 0x20, 0x01,
+        0x42, 0x11, 0x10, 0x04, 0x21, 0x01, 0x0c, 0x01, 0x05, 0x20, 0x01, 0x02, 0x7e, 0x23, 0x00,
+        0x41, 0x10, 0x6b, 0x22, 0x03, 0x24, 0x00, 0x03, 0x40, 0x02, 0x40, 0x02, 0x40, 0x20, 0x03,
+        0x02, 0x7f, 0x20, 0x04, 0x41, 0x07, 0x46, 0x04, 0x40, 0x20, 0x03, 0x41, 0x08, 0x6a, 0x20,
+        0x02, 0x42, 0x04, 0x86, 0x42, 0x09, 0x84, 0x37, 0x03, 0x00, 0x41, 0x00, 0x0c, 0x01, 0x0b,
+        0x20, 0x04, 0x41, 0x0a, 0x47, 0x04, 0x40, 0x42, 0x01, 0x21, 0x01, 0x20, 0x04, 0x41, 0x80,
+        0x80, 0x40, 0x6b, 0x2d, 0x00, 0x00, 0x22, 0x05, 0x41, 0xdf, 0x00, 0x46, 0x0d, 0x02, 0x20,
+        0x05, 0xad, 0x21, 0x01, 0x02, 0x40, 0x02, 0x40, 0x20, 0x05, 0x41, 0x30, 0x6b, 0x41, 0xff,
+        0x01, 0x71, 0x41, 0x0a, 0x4f, 0x04, 0x40, 0x20, 0x05, 0x41, 0xc1, 0x00, 0x6b, 0x41, 0xff,
+        0x01, 0x71, 0x41, 0x1a, 0x49, 0x0d, 0x01, 0x20, 0x05, 0x41, 0xe1, 0x00, 0x6b, 0x41, 0xff,
+        0x01, 0x71, 0x41, 0x1a, 0x49, 0x0d, 0x02, 0x20, 0x03, 0x41, 0x01, 0x36, 0x02, 0x04, 0x20,
+        0x03, 0x41, 0x08, 0x6a, 0x20, 0x05, 0x36, 0x02, 0x00, 0x41, 0x01, 0x0c, 0x04, 0x0b, 0x20,
+        0x01, 0x42, 0x2e, 0x7d, 0x21, 0x01, 0x0c, 0x04, 0x0b, 0x20, 0x01, 0x42, 0x35, 0x7d, 0x21,
+        0x01, 0x0c, 0x03, 0x0b, 0x20, 0x01, 0x42, 0x3b, 0x7d, 0x21, 0x01, 0x0c, 0x02, 0x0b, 0x20,
+        0x03, 0x41, 0x00, 0x36, 0x02, 0x04, 0x20, 0x03, 0x41, 0x08, 0x6a, 0x41, 0x07, 0x36, 0x02,
+        0x00, 0x41, 0x01, 0x0b, 0x36, 0x02, 0x00, 0x0c, 0x01, 0x0b, 0x20, 0x04, 0x41, 0x01, 0x6a,
+        0x21, 0x04, 0x20, 0x01, 0x20, 0x02, 0x42, 0x06, 0x86, 0x84, 0x21, 0x02, 0x0c, 0x01, 0x0b,
+        0x0b, 0x20, 0x03, 0x28, 0x02, 0x00, 0x45, 0x04, 0x40, 0x20, 0x03, 0x29, 0x03, 0x08, 0x20,
+        0x03, 0x41, 0x10, 0x6a, 0x24, 0x00, 0x0c, 0x01, 0x0b, 0x00, 0x0b, 0x10, 0x00, 0x20, 0x00,
+        0x10, 0x01, 0x10, 0x02, 0x20, 0x06, 0x41, 0x10, 0x6a, 0x24, 0x00, 0x0b, 0x0b, 0x0b, 0x0b,
+        0x10, 0x01, 0x00, 0x41, 0x80, 0x80, 0xc0, 0x00, 0x0b, 0x07, 0x76, 0x65, 0x63, 0x5f, 0x65,
+        0x72, 0x72,
+    ];
+    let scob0 = ScObject::Binary(code0.try_into().unwrap());
+    let val0 = ScVal::Object(Some(scob0));
+    let le0 = LedgerEntry {
+        last_modified_ledger_seq: 0,
+        data: LedgerEntryData::ContractData(ContractDataEntry {
+            contract_id: id0,
+            key: key.clone(),
+            val: val0,
+        }),
+        ext: LedgerEntryExt::V0,
+    };
+
+    // 2nd level, the guest contract
+    let id1: Hash = [1; 32].into();
+    let storage_key1 = LedgerKey::ContractData(LedgerKeyContractData {
+        contract_id: id1.clone(),
+        key: key.clone(),
+    });
+    let code1: [u8; 170] = [
+        0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00, 0x01, 0x11, 0x03, 0x60, 0x00, 0x01, 0x7e,
+        0x60, 0x03, 0x7e, 0x7e, 0x7e, 0x01, 0x7e, 0x60, 0x01, 0x7e, 0x01, 0x7e, 0x02, 0x0d, 0x02,
+        0x01, 0x76, 0x01, 0x5f, 0x00, 0x00, 0x01, 0x76, 0x01, 0x38, 0x00, 0x01, 0x03, 0x02, 0x01,
+        0x02, 0x05, 0x03, 0x01, 0x00, 0x10, 0x06, 0x19, 0x03, 0x7f, 0x01, 0x41, 0x80, 0x80, 0xc0,
+        0x00, 0x0b, 0x7f, 0x00, 0x41, 0x80, 0x80, 0xc0, 0x00, 0x0b, 0x7f, 0x00, 0x41, 0x80, 0x80,
+        0xc0, 0x00, 0x0b, 0x07, 0x2f, 0x04, 0x06, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x02, 0x00,
+        0x07, 0x76, 0x65, 0x63, 0x5f, 0x65, 0x72, 0x72, 0x00, 0x02, 0x0a, 0x5f, 0x5f, 0x64, 0x61,
+        0x74, 0x61, 0x5f, 0x65, 0x6e, 0x64, 0x03, 0x01, 0x0b, 0x5f, 0x5f, 0x68, 0x65, 0x61, 0x70,
+        0x5f, 0x62, 0x61, 0x73, 0x65, 0x03, 0x02, 0x0a, 0x29, 0x01, 0x27, 0x01, 0x01, 0x7f, 0x23,
+        0x00, 0x41, 0x10, 0x6b, 0x22, 0x01, 0x24, 0x00, 0x20, 0x01, 0x10, 0x00, 0x42, 0xd1, 0x00,
+        0x20, 0x00, 0x10, 0x01, 0x37, 0x03, 0x00, 0x20, 0x01, 0x29, 0x03, 0x00, 0x20, 0x01, 0x41,
+        0x10, 0x6a, 0x24, 0x00, 0x0b,
+    ];
+    let scob1 = ScObject::Binary(code1.try_into().unwrap());
+    let val1 = ScVal::Object(Some(scob1));
+    let le1 = LedgerEntry {
+        last_modified_ledger_seq: 0,
+        data: LedgerEntryData::ContractData(ContractDataEntry {
+            contract_id: id1,
+            key,
+            val: val1,
+        }),
+        ext: LedgerEntryExt::V0,
+    };
+
+    // create storage map and footprint
+    let mut map = OrdMap::unit(storage_key0.clone(), Some(le0));
+    map.insert(storage_key1.clone(), Some(le1));
+    let mut footprint = Footprint::default();
+    footprint.record_access(&storage_key0, AccessType::ReadOnly);
+    footprint.record_access(&storage_key1, AccessType::ReadOnly);
+
+    // initialize storage and host
+    let storage = Storage::with_enforcing_footprint_and_map(footprint, map);
+    let host = Host::with_storage(storage);
+    // prepare arguments
+    let obj = host.test_bin_obj(&[0; 32])?;
+    let sym = Symbol::from_str("del_call");
+    let args = host.test_vec_obj::<i32>(&[1])?;
+    // try call
+    let sv = host.try_call(obj.to_object(), sym.into(), args.clone().into())?;
+    let code = ScHostObjErrorCode::VecIndexOutOfBound;
+    let exp_st: Status = code.into();
+    assert_eq!(sv.get_payload(), exp_st.as_ref().get_payload());
+    // call
+    let res = host.call(obj.to_object(), sym.into(), args.into());
+    assert!(HostError::result_matches_err_status(res, code));
+    Ok(())
+}
