+++ stellar-contract-env-host/src/test/map.rs
@@ -1,0 +1,191 @@
+use crate::{
+    xdr::{ScMap, ScMapEntry, ScObject, ScVal, ScVec},
+    CheckedEnv, Host, HostError, RawVal, RawValConvertible, Symbol, UNKNOWN_ERROR,
+};
+
+#[test]
+fn map_put_has_and_get() -> Result<(), HostError> {
+    let host = Host::default();
+    let scmap: ScMap = host.map_err(
+        vec![
+            ScMapEntry {
+                key: ScVal::U32(1),
+                val: ScVal::U32(2),
+            },
+            ScMapEntry {
+                key: ScVal::U32(2),
+                val: ScVal::U32(4),
+            },
+        ]
+        .try_into(),
+    )?;
+    let scobj = ScObject::Map(scmap);
+    let obj = host.to_host_obj(&scobj)?;
+    let k: RawVal = 3_u32.into();
+    let v: RawVal = 6_u32.into();
+    assert!(!bool::try_from(host.map_has(obj.to_object(), k)?)?);
+    let obj1 = host.map_put(obj.to_object(), k, v)?;
+    assert!(bool::try_from(host.map_has(obj1, k)?)?);
+    let rv = host.map_get(obj1, k)?;
+    let v = unsafe { <u32 as RawValConvertible>::unchecked_from_val(rv) };
+    assert_eq!(v, 6);
+    Ok(())
+}
+
+#[test]
+fn map_prev_and_next() -> Result<(), HostError> {
+    let host = Host::default();
+    let scmap: ScMap = host.map_err(
+        vec![
+            ScMapEntry {
+                key: ScVal::U32(1),
+                val: ScVal::U32(2),
+            },
+            ScMapEntry {
+                key: ScVal::U32(4),
+                val: ScVal::U32(8),
+            },
+        ]
+        .try_into(),
+    )?;
+    let scobj = ScObject::Map(scmap);
+    let obj = host.to_host_obj(&scobj)?;
+    // prev
+    {
+        assert_eq!(
+            host.map_prev_key(obj.to_object(), 0_u32.into())?
+                .get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(obj.to_object(), 1_u32.into())?
+                .get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(obj.to_object(), 2_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(1).get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(obj.to_object(), 4_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(1).get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(obj.to_object(), 5_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(4).get_payload()
+        );
+    }
+    // next
+    {
+        assert_eq!(
+            host.map_next_key(obj.to_object(), 5_u32.into())?
+                .get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(obj.to_object(), 4_u32.into())?
+                .get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(obj.to_object(), 3_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(4).get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(obj.to_object(), 1_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(4).get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(obj.to_object(), 0_u32.into())?
+                .get_payload(),
+            RawVal::from_u32(1).get_payload()
+        );
+    }
+    Ok(())
+}
+
+#[test]
+fn map_prev_and_next_heterogeneous() -> Result<(), HostError> {
+    let host = Host::default();
+    let scmap: ScMap = host.map_err(
+        vec![ScMapEntry {
+            key: ScVal::U32(1),
+            val: ScVal::U32(2),
+        }]
+        .try_into(),
+    )?;
+    let scvec: ScVec = host.test_scvec::<u32>(&[1])?;
+
+    let scobj_map = ScObject::Map(scmap);
+    let scobj_vec = ScObject::Vec(scvec);
+    let sym = Symbol::from_str("symbol");
+
+    let obj_map = host.to_host_obj(&scobj_map)?;
+    let obj_vec = host.to_host_obj(&scobj_vec)?;
+
+    let mut test_map = host.map_new()?;
+    test_map = host.map_put(test_map, 2u32.into(), 4u32.into())?;
+    test_map = host.map_put(test_map, obj_map.clone().into(), 4u32.into())?;
+    test_map = host.map_put(test_map, obj_vec.clone().into(), 4u32.into())?;
+    test_map = host.map_put(test_map, sym.clone().into(), 4u32.into())?;
+    // The key ordering should be [u32, vec, map, symbol]
+    // prev
+    {
+        assert_eq!(
+            host.map_prev_key(test_map, 0_u32.into())?.get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(test_map, 4_u32.into())?.get_payload(),
+            RawVal::from_u32(2).get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(test_map, obj_vec.clone().into())?
+                .get_payload(),
+            RawVal::from_u32(2).get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(test_map, obj_map.clone().into())?
+                .get_payload(),
+            obj_vec.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_prev_key(test_map, sym.clone().into())?
+                .get_payload(),
+            obj_map.to_raw().get_payload()
+        );
+    }
+    // next
+    {
+        assert_eq!(
+            host.map_next_key(test_map, 0_u32.into())?.get_payload(),
+            RawVal::from_u32(2).get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(test_map, 4_u32.into())?.get_payload(),
+            obj_vec.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(test_map, obj_vec.clone().into())?
+                .get_payload(),
+            obj_map.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(test_map, obj_map.clone().into())?
+                .get_payload(),
+            sym.to_raw().get_payload()
+        );
+        assert_eq!(
+            host.map_next_key(test_map, sym.clone().into())?
+                .get_payload(),
+            UNKNOWN_ERROR.to_raw().get_payload()
+        );
+    }
+
+    Ok(())
+}
